# docs/swagger.yaml
definitions:
  handler.ContainerInfo:
    properties:
      created_at:
        type: string
      id:
        type: string
      image:
        type: string
      name:
        type: string
      state:
        type: string
      status:
        type: string
    type: object
  images.Image:
    properties:
      created_at:
        example: "2013-08-20T18:52:09.000Z"
        type: string
      id:
        example: a407dee395ed
        type: string
      image_name:
        example: dockware/dev
        type: string
      image_tag:
        example: 6.6.8.2
        type: string
      size:
        example: 1048576
        type: integer
    type: object
  images.ImageDeleteResponse:
    properties:
      message:
        example: Docker Image removed successfully
        type: string
      status:
        example: success
        type: string
    type: object
  images.PullImageRequest:
    properties:
      image_name:
        example: dockware/dev
        type: string
      image_tag:
        example: 6.6.8.2
        type: string
    type: object
  images.PullImageResponse:
    properties:
      message:
        example: Docker Image created successfully
        type: string
      status:
        example: success
        type: string
    type: object
info:
  contact: {}
paths:
  /api/health:
    get:
      consumes:
      - '*/*'
      description: Get the status of server.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      summary: Show the status of server.
      tags:
      - System Management
  /api/images:
    get:
      consumes:
      - application/json
      description: Get a list of docker image
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/images.Image'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List of Docker Image
      tags:
      - Docker Image Management
    post:
      consumes:
      - application/json
      description: Pulls a docker image and register it into the system
      parameters:
      - description: Image Input
        in: body
        name: image
        required: true
        schema:
          $ref: '#/definitions/images.PullImageRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/images.PullImageResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Pull Docker Image
      tags:
      - Docker Image Management
  /api/images/{id}:
    delete:
      consumes:
      - application/json
      description: Removes a docker image form the system
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/images.ImageDeleteResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove Docker Image
      tags:
      - Docker Image Management
    get:
      consumes:
      - application/json
      description: Get details about a docker image
      parameters:
      - description: Image ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/images.Image'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Details about a Docker Image
      tags:
      - Docker Image Management
  /api/sandboxes:
    get:
      consumes:
      - application/json
      description: List sandbox environments
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handler.ContainerInfo'
      summary: List sandbox environments
      tags:
      - Sandbox Management
swagger: "2.0"